chore: pascal-case renames
