+++ src/module/actor/data-model-classes/data-model-character-encumbrance.js
